BackEnd/requirements.txt
django>=5.2,<6
djangorestframework
django-cors-headers
djangorestframework-simplejwt
requests
